requirements.txt
dj-database-url == 0.2.1
django == 1.4.2
django-braces == 0.1.7
django-configurations == 0.1
django-model-utils == 1.1.0
django-selectable == 0.6.1
django-taggit == 0.9.3
django-tastypie == 0.9.11
gunicorn == 0.15.0
psycopg2 == 2.4.5
